services/airflow-app/requirements.txt
apache-airflow>=2.10.5
apache-airflow-providers-apache-spark>=5.0.0
psycopg2-binary>=2.9.10
pyspark>=3.5.5
confluent-kafka>=2.8.2
